+++ src/craftground/MinecraftEnv/src/main/java/com/kyhsgeekcode/minecraftenv/customentity/RealisticHumanModel.kt
@@ -1,4 +1,4 @@
-package com.kyhsgeekcode.minecraftenv.custom_entity
+package com.kyhsgeekcode.minecraftenv.customentity
 
 import net.minecraft.client.model.ModelData
 import net.minecraft.client.model.ModelPart
